gcovFlushWithProp: Add new static lib
BZ: 81124

To use gcov on a shared lib in an embedded platform,
we need to force gcov flush manually.
This new static lib starts a new thread
to detect the Android property "gcov.flush.force" value.
When this property is "true", this thread forces a gcov flush.
Add the Doxygen config file and tag "html" folder into git ignore

Change-Id: Ib2fd3c971fc9c4e7967ba3117e885c3da20fbe5e
Signed-off-by: Jeremie Graulle <jeremiex.graulle@intel.com>
Reviewed-on: http://android.intel.com:8080/90305
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Gonzalve, Sebastien <sebastien.gonzalve@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>