Utilities: Add a header to convert string safely.
BZ: 105037

In many places in code we need to convert strings to int, uint, double etc. but
we did not have any conversion utility function to perform conversion safely.
This leads to many incorrect uses of atoi() or strtoX() call without correct
check.

This patch adds a header to provide a conversion function which allows to
perform those conversions safely.

Change-Id: I0d522b315cfafca5c2e54847db1e206f51573134
Signed-off-by: Sebastien Gonzalve <sebastien.gonzalve@intel.com>
Reviewed-on: http://android.intel.com:8080/105752
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Reviewed-by: Graulle, JeremieX <jeremiex.graulle@intel.com>
Reviewed-by: Rocard, KevinX <kevinx.rocard@intel.com>
Reviewed-by: Cahuzac, ThomasX <thomasx.cahuzac@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>