utilities: add a Dynamic Library loader class
In several places in the audio comms code, we use the dynamic loading of
libraries (use dlopen and dlsym). For now, code is duplicated and the dlclose
is never done.

In order to have proper use of dynamic libraries, this patch adds a dynamic
library wrapper that would provide a suitable API for C++ use.

Change-Id: I5177b5e79e71c0b6c9b5972864cc8c810f17bf4c
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-9102
Signed-off-by: Sebastien Gonzalve <sebastien.gonzalve@intel.com>
Reviewed-on: https://android.intel.com:443/300887