Add an optional concept for deserialization
BZ: 211893

The serializer used to fail if an element was not present during
deserialization. As a result an empty string will lead to a deserialization
failure.

Add an optional flag in Child to indicate that it is optional. In this
case, the default object (empty constructor) will be used to set the
parent.

Change-Id: I394a39ad2141f46521ffde6414f92ed1effa1ddd
Origin-Change-Id: I9ec89a6037cbfe6da922e0e9fe3083fffab5e44e
Signed-off-by: Kevin Rocard <kevinx.rocard@intel.com>
Reviewed-on: 155483