Risk of bug in libfull_rw static library
BZ: 168078

Functions full_read() and full_write() read() and write() until the
desired amount of data is effectively read or written using standard
system calls and handling EAGAIN and EINTR errno.
For that reason, they include a loop with a data counter. The issue
is that the data counter may be wrong in case the EAGAIN or EINTR
errno is repported, because both functions do not check the system
call to read() or write() has return a positive value before to add
it to the data counter.

This patch updates the functions full_read() and full_write() to
ensure they update their data counters only if the system call has
returned a positive value.

Change-Id: I58caba2e6dde05df78d977ece54d1a9623cb4353
Signed-off-by: Mathieu Soulard <mathieu.soulard@intel.com>