Checkfiles: quiet unwanted error message
BZ: 177489

Checkfile script of all AudioComms projects execute a "base" script
located in a private repo. When executed in the context of an external
delivery, this redirection fails and the script displays an unexpected
error message.

This patch prevents error message from being displayed and replaces it
by an info message.

Change-Id: I4e372ab7712a5df60374cd33ec13a5df787b15ee
Signed-off-by: Renaud de Chivre <renaud.de.chivre@intel.com>