convert.hpp: specialize convertTo for float and double
BZ: 176178

The stlport behaviour differs from the GNU STL regarding overflows: instead of
failing to convert, it converts values larger than max() to Infinity.

The specialization takes this into account and returns an error if the
converted value is +/-Infinity.  It also returns an error if a "NaN" (Not A
Number) is returned by the STL.

Change-Id: I4ecc21aa109dffc2867d4e3b296181bd287df495
Signed-off-by: David Wagner <david.wagner@intel.com>