Thread now call process() at least once
Regardless of the start stop delay, if the process() function (1) can
be called zero time (2) it can also be called once (3).

As the zero and 1 executions cases are undetermined, change the while
loop in a do while loop. With this change process() will always be
called once.

(1): the thread "do the job" function
(2): because the thread was not scheduled enough between start and stop
(3): if only the if(not stopRequested) as scheduled in loop()

Change-Id: I12b409fa435763940b61e101b3f0fae08478a85b
Tracked-On: IMINAN-24023
Signed-off-by: Kevin Rocard <kevinx.rocard@intel.com>
Reviewed-on: 178112
Reviewed-on: 337648