[WORKAROUND] Moving remote parameter plugin to parameter-framework folder
BZ: 184301

The remote parameter plugin makes the external build fail because it is
located in a public folder (currently all the parameter framework
plugins must be private).

This patch removes the plugin from the utilities project, it will be
added into the parameter-framework folder (which is private)
in a subsequent patch.

This patch is workaround because the definitive plugin
location will be known when the decision "parameter framework plugins
can be public or not" will be made.

Change-Id: I8f12250d69f312991f31a6091ae0ce553d774e53
Signed-off-by: Thomas Cahuzac <thomasx.cahuzac@intel.com>