Use FirstChildElement
BZ: 211893

The xml serializor framework was using tinyXml's FirstChild(tag)
and expecting to get a xml element with the requested tag.
However First child returns the first child with the matching 'value'.
Not specifically an xml element, it could be a XmlText.

Use FirstChildElement in order to be sure to get an XmlElement.

Change-Id: Ia90dd0a96a13e13b01ec9feab61a32fc2a7fdf07
Origin-Change-Id: I05f13705b6b17b6aa1909f4235f63e7a6e1693ec
Signed-off-by: Kevin Rocard <kevinx.rocard@intel.com>
Reviewed-on: 190580