Avoid leak in polymorphic deserialization
BZ: 211893

When deserializing a polymorphic pointer, the framework allocates
the matching derived class and return a pointer to Base.
Deletion was not handled by the framework and thus pointee were
leaked if the setter did not take the pointee ownership.

Change the Child::takeOwnership meaning for polymorphic trait.
takeOwnership to true now mean that the setter takes the ownership of
the pointee not the pointer.

Change-Id: Ib0fb6e6d2d3e87e3cf20cd29f01fd2445031f947
Origin-Change-Id: Ibbe3ae09973d3fefbfa57a24d1344d1cc114f424
Signed-off-by: Kevin Rocard <kevinx.rocard@intel.com>
Reviewed-on: 192545