result is now moved out from ACME
BZ: 211893

The result library is only used in ACME but could be usefull for at
least other audio comms components.

For that reason it has been moved out from the ACME and is now known
as libacresult, located in audiocomms utility project.

This patch removes it from ACME and update ACME to build using the
'new' libacresult instead of libcmeresult.

Change-Id: Ib5841e6112369bb8395ecd39b9648184956280ac
Origin-Change-Id: I1d5524c9a4d864a26600e8a305ae629186462aef
Signed-off-by: Mathieu Soulard <mathieu.soulard@intel.com>