Result may no longer be casted to error code
BZ: 204605

Implicit casting from Result to Code hided bugs where a result was
implicitly casted to a bool through Code (Result=>Code(enum)=>int=>bool).
As Code is an enum, it makes no sense to cast it to bool,
but this can not be forbidden (legacy C inheritance).
Typical hided bugs:
    if(result) // Wrong, but compiled
    if(result.success()) // Wrong, but compiled
    bool f() { return g(); } // Wrong if g returns a Result

Remove the cast operator and add == and != operator to continue
testing a Result again a Code easily.

Change-Id: Ibe77cf472b7c7bf0d8ac74b38b4776fb8d533800
Signed-off-by: Kevin Rocard <kevinx.rocard@intel.com>
Reviewed-on: https://android.intel.com/208321
Reviewed-by: Soulard, Mathieu <mathieu.soulard@intel.com>