[AUDIO] Rename gcov flush host utility library
BZ: 202797

This patch introduces the gcov flush utility library on target with the
same name as used on host in order to use the quality hook
regardless of the type of build (host / target).
It keeps the flush version using property only on target as property lib
dependancy is not available on host.

Change-Id: If5c88da470b4559f0e9b5fd9d2185e62b191c1ad
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>