XML serializer can get takes a config param
BZ: 212064

XML serializer was getting it's configuration from template specialization
of an hard coded class.This configuration is a class that given a class
as template parameter returns the corresponding trait.
As template need to be specialized in their namespace,
having the configuration part of the utilities::serializer namespace
makes specialization very verbose.
The other problem is that this configuration is a kind of compile time global variable.

Give the possibility to provide configuration as template parameter.
Keep old hard coded config class for retrocompatibitily.

Change-Id: I49c0add7d8dfd966ef4dbf0f89c4578470ce58fc
Signed-off-by: Kevin Rocard <kevinx.rocard@intel.com>
Reviewed-on: 211780
Reviewed-by: Gonzalve, Sebastien <sebastien.gonzalve@intel.com>
Tested-by: Gonzalve, Sebastien <sebastien.gonzalve@intel.com>