Change Result format
BZ: 211947

Result format was as such:
Code <error code> (<literal code>): <error message>
This formating makes the error code important as it's the first item,
although is is usually of little use.
The folowing code:
    result1 & (result2 << message) & (result3 << subResult1 << subResult2)
used to format as such:
    Code <error code1>: ( && \
    Code <error code2>: <literal code2> (<error message2> && \
    Code <error code3>: <literal code3> \
        (<formated sub result1> : <formated subResult2>)))

Change the format to:
    <literal code1> (Code <error code1>) && \
    <literal code2> (Code <error code2>): <error message2> && \
    <literal code3> (Code <error code3>): \
        {<formated sub result1> && <formated sub result2>}

The literal version of the error code comes before it's numeric version.
Result from the same level are separeted by && and are not in a
bracket pair ().
Results from a lower level are surounded by braces {}.

Change-Id: Ic701f76d8d86bd6e4f2ffe5fc1d33e064f4b0219
Signed-off-by: Kevin Rocard <kevinx.rocard@intel.com>
Reviewed-on: https://android.intel.com/220161