ScopedTrace: Avoid useless processing
BZ: 225404

When ScopedTrace is present in a software program
a system call logs are displayed during profiling (i.e: strace)
even if the trace logs are not activated.

This patch aims to make the system call only if the trace logs are
activated so that the do not pollute profiling.

Change-Id: I812be3f23576bcf589945704f5ca1528dcc1966f
Signed-off-by: Michael Soares <michaelx.soares@intel.com>
Reviewed-on: 235241
Reviewed-by: Gonzalve, Sebastien <sebastien.gonzalve@intel.com>
Tested-by: Gonzalve, Sebastien <sebastien.gonzalve@intel.com>
Reviewed-on: 282963