Add SignalProcessing lib Unit Tests
Unit tests for Signal Processing lib has been added.

As Signal Processing lib is template, TypedTest class
has been added. It's a template class which allows to
write a template test function which will be launch with
different types given in a TypeList.
This class is used with gtest and works like the
TYPED_TEST macro which is unavailable with our
buildsystem.

Change-Id: I244a85cf88f261e8660fa7c5dfc2ca4db5583d40
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-441
Signed-off-by: Jules Clero <jules.clero@intel.com>
Reviewed-on: https://android.intel.com:443/245485