Fix SignalProcessing algorithm
BZ: 225404

Unit Tests implementation has shown errors in the
Cross-Correlation algorithm.

The mean calculation has been fixed.

The cross-correlation function can now return a
non success result value if we launch the algorithm
on a constant signal. Indeed, the algorithm is
undefined when a constant signal is compared to
a non-constant one.

Change-Id: Iad9544ebbfcd4ac8b39a19805923e5ca145fcc3a
Signed-off-by: Jules Clero <jules.clero@intel.com>
Reviewed-on: 247991
Reviewed-by: Gonzalve, Sebastien <sebastien.gonzalve@intel.com>
Tested-by: Gonzalve, Sebastien <sebastien.gonzalve@intel.com>
Reviewed-on: 282965