Refactor SignalProcessing lib
The SignalProcessing lib was used to analyse audio files.
The memory mapping of the audio file has been moved in the
FileMapper class.
This class allows to map any file in memory. This file
can then be accessed through a pointer.

Now, when we need to analyse an audio file with the
SignalProcessing lib, users will have to map the file
in memory using FileMapper.

Change-Id: Ia3ea57e83afd8d1de1eca0ea7d6edf2d42d0b3db
Tracked-On: IMINAN-441
Signed-off-by: Jules Clero <jules.clero@intel.com>
Reviewed-on: 245486