Fix Semaphore wait with timeout function
BZ: 222915

The wait with timeout function should return false
if a timeout occurs.

As the errno value is not erased between two calls to
wait if no error occurs. It is possible that the function
return false even if no timeout occurs.

Indeed, the return value should not be the comparison of
errno and ETIMEDOUT. It should be the comparison of the call
to the posix wait API and 0.

This patch corrects the return value of the wait method.

Change-Id: I60f8c005677baf874124d1986b602ee00ea1a0d2
Signed-off-by: Jules Clero <jules.clero@intel.com>
Reviewed-on: 258014