Remove LOCAL_IMPORT_C_INCLUDE_DIRS_FROM_STATIC_LIBRARIES
This variable has been added by us in the build system but it isn't part of the
AOSP build system.  It was introduced, as its name suggests, to get the
needed include directories for a given static library without linking to it.

Instead, this patch removes this non-upstreamable feature and link
directly against the static library instead. It will not impact the code size.
This patch created the target libproperty_includes to import the headers.
The recommended include is also used to import STL lib.

Change-Id: I0f7cdf328ee082fa5b41e9fc3171e626ebd3eb99
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-1808
Signed-off-by: pafonsoX <philippex.afonso@intel.com>
Reviewed-on: https://android.intel.com:443/259905