Add UniquePtr void* comparison operators
A unique ptr could not be compared to a raw pointer.
This was unwanted for multiple reasons:
 - behavior differ from a raw pointer.
 - could not compare to NULL

It should be fixed by adding comparison operators
to nullptr_t but it is only introduced in c++11.

Fix this by adding comparison operators to void*.
This permits the comparison to NULL but unfortunately
also to an other pointer of different type.
Nevertheless it is considered a minor cost compared to
the ability to compare with NULL.

Also remove the isNull method is it is now useless
(and nobody used it) and is moves away from the raw
pointer usage.

Issue: GMINL-5664

Change-Id: I8a2426fef8bc451f1e77eca7614014fb1d81517b
Signed-off-by: Kevin Rocard <kevinx.rocard@intel.com>
Reviewed-on: https://android.intel.com/327073
Reviewed-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Tested-by: Esclafer, CatherineX <catherinex.esclafer@intel.com>