Fix stlport usage in interface provider makefile
According to libstlport.mk, to use stlport for a given target, we shall
"include external/stlport/libstlport.mk" in the target.
This was not done in the interface provider makefiles.

This patch change all targets related to the interface provider
which require stlport to include it in the correct way.

Change-Id: I293f44f07ad2175190bdd2eaaa150001ec571a00
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-10871
Signed-off-by: Mattijs Korpershoek <mattijsx.korpershoek@intel.com>
Reviewed-on: https://android.intel.com:443/306040